move Head to where it is used
