+++ src/home/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import Head from '../component/Head';
+import Head from './Head';
 
 const Home = (): JSX.Element => {
   const { t } = useTranslation();
